Pull recipe processing out into a separate module
