+++ meal_planner/main.py
@@ -22,6 +22,7 @@ from starlette.staticfiles import StaticFiles
 
 from meal_planner.api.recipes import API_ROUTER as RECIPES_API_ROUTER
 from meal_planner.models import RecipeBase
+from meal_planner.services.recipe_processing import postprocess_recipe
 
 MODEL_NAME = "gemini-2.0-flash"
 ACTIVE_RECIPE_EXTRACTION_PROMPT_FILE = "20250505_213551__terminal_periods_wording.txt"
@@ -507,7 +508,7 @@ async def fetch_and_clean_text_from_url(recipe_url: str) -> str:
 
 async def extract_recipe_from_text(page_text: str) -> RecipeBase:
     """Extracts and post-processes a recipe from text."""
-    logger.info("Starting recipe extraction from text:\n%s", page_text)
+    logger.info("Starting recipe extraction from text:\\n%s", page_text)
     try:
         prompt_file_path = _get_prompt_path(
             "recipe_extraction", ACTIVE_RECIPE_EXTRACTION_PROMPT_FILE
@@ -543,58 +544,6 @@ async def extract_recipe_from_url(recipe_url: str) -> RecipeBase:
     return await extract_recipe_from_text(
         await fetch_and_clean_text_from_url(recipe_url)
     )
-
-
-def postprocess_recipe(recipe: RecipeBase) -> RecipeBase:
-    """Post-processes the extracted recipe data."""
-    if recipe.name:
-        recipe.name = _postprocess_recipe_name(recipe.name)
-    if recipe.ingredients:
-        recipe.ingredients = [
-            _postprocess_ingredient(i) for i in recipe.ingredients if i.strip()
-        ]
-    if recipe.instructions:
-        recipe.instructions = [
-            _postprocess_instruction(i) for i in recipe.instructions if i.strip()
-        ]
-
-    return recipe
-
-
-def _postprocess_recipe_name(name: str) -> str:
-    """Cleans and standardizes the recipe name."""
-    return _close_parenthesis(
-        name.strip().removesuffix("recipe").removesuffix("Recipe").title().strip()
-    )
-
-
-def _postprocess_ingredient(ingredient: str) -> str:
-    """Cleans and standardizes a single ingredient string."""
-    return _close_parenthesis(" ".join(ingredient.split()).strip().replace(" ,", ","))
-
-
-def _postprocess_instruction(instruction: str) -> str:
-    """Cleans and standardizes a single instruction string."""
-    return (
-        _remove_leading_step_numbers(html.unescape(instruction))
-        .replace(" ,", ",")
-        .replace(" ;", ";")
-        .strip()
-    )
-
-
-def _remove_leading_step_numbers(instruction: str) -> str:
-    """Removes leading step numbers like "Step 1", "Step 1:", "1.", "1 " """
-    return re.sub(
-        r"^\s*(?:Step\s*\d+|\d+)\s*[:.]?\s*", "", instruction, flags=re.IGNORECASE
-    ).strip()
-
-
-def _close_parenthesis(text: str) -> str:
-    """Appends a closing parenthesis if an opening one exists without a closing one."""
-    if "(" in text and ")" not in text:
-        return text + ")"
-    return text
 
 
 CSS_ERROR_CLASS = f"{TextT.error} mb-4"
