+++ tests/services/test_recipe_processing.py
@@ -1,0 +1,75 @@
+import pytest
+
+from meal_planner.models import RecipeBase
+from meal_planner.services.recipe_processing import postprocess_recipe
+
+
+class TestPostprocessRecipeName:
+    DUMMY_INGREDIENTS = ["dummy ingredient"]
+    DUMMY_INSTRUCTIONS = ["dummy instruction"]
+
+    @pytest.mark.parametrize(
+        "input_name, expected_name",
+        [
+            ("  my awesome cake  ", "My Awesome Cake"),
+            ("Another Example recipe ", "Another Example"),
+            ("Another Example Recipe ", "Another Example"),
+            ("Recipe (unclosed", "Recipe (Unclosed)"),
+        ],
+    )
+    def test_postprocess_recipe_name(self, input_name: str, expected_name: str):
+        input_recipe = RecipeBase(
+            name=input_name,
+            ingredients=self.DUMMY_INGREDIENTS,
+            instructions=self.DUMMY_INSTRUCTIONS,
+        )
+        processed_recipe = postprocess_recipe(input_recipe)
+        assert processed_recipe.name == expected_name
+
+
+class TestPostprocessRecipe:
+    def test_postprocess_ingredients(self):
+        recipe = RecipeBase(
+            name="Test Name",
+            ingredients=[
+                "  Ingredient 1 ",
+                "Ingredient 2 (with parens)",
+                " Ingredient 3, needs trim",
+                "  ",
+                "Multiple   spaces",
+                " Ends with comma ,",
+            ],
+            instructions=["Step 1"],
+        )
+        processed = postprocess_recipe(recipe)
+        assert processed.ingredients == [
+            "Ingredient 1",
+            "Ingredient 2 (with parens)",
+            "Ingredient 3, needs trim",
+            "Multiple spaces",
+            "Ends with comma,",
+        ]
+
+    def test_postprocess_instructions(self):
+        recipe = RecipeBase(
+            name="Test Name",
+            ingredients=["Ing 1"],
+            instructions=[
+                " Step 1: Basic step. ",
+                "2. Step with number.",
+                "  Step 3 Another step.",
+                "No number.",
+                "  ",
+                " Ends with semicolon ;",
+                " Has comma , in middle",
+            ],
+        )
+        processed = postprocess_recipe(recipe)
+        assert processed.instructions == [
+            "Basic step.",
+            "Step with number.",
+            "Another step.",
+            "No number.",
+            "Ends with semicolon;",
+            "Has comma, in middle",
+        ]
